requirements.txt
psycopg2-binary
pillow
python-dotenv
pytest